backend/requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
Authlib==1.6.6
bcrypt==5.0.0
billiard==4.2.4
blis==1.3.3
catalogue==2.0.10
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.4
cymem==2.0.13
deprecation==2.1.0
ecdsa==0.19.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.128.0
filelock==3.20.3
fsspec==2026.1.0
greenlet==3.3.1
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.5
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
kombu==5.6.2
loguru==0.7.3
MarkupSafe==3.0.3
mpmath==1.3.0
murmurhash==1.0.15
networkx==3.6.1
numpy==2.4.1
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
preshed==3.0.12
prompt_toolkit==3.0.52
protobuf==6.33.5
psycopg-binary==3.2.9
psycopg==3.2.9
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.0
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
redis==7.1.0
regex==2026.1.15
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
sentencepiece==0.2.1
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.46
srsly==2.5.2
starlette==0.50.0
sympy==1.14.0
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.2
transformers==5.0.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
validators==0.35.0
vine==5.1.0
wasabi==1.1.3
wcwidth==0.5.3
weasel==0.4.3
weaviate-client==4.19.2
win32_setctime==1.2.0
wrapt==2.0.1
PyPDF2==3.0.1
python-docx==1.1.2
pgvector==0.3.6
